perf: Better account rebuild performance
This revision updates the account rebuild process to improve performance.
The code was refactored and logic was moved into using defaultdict and bulk_create methods which significantly increases the speed and efficiency.
It accomplishes this by updating all account transactions by date and by adding and saving all account balances and histories in bulk.